pyproject.toml
[project]
name = "banking-project"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
	"asyncpg>=0.30.0",
	"dash>=3.0.4",
	"flask-sqlalchemy>=3.1.1",
	"gunicorn>=23.0.0",
	"openpyxl>=3.1.5",
	"pandas>=2.2.3",
	"psycopg>=3.2.7",
	"psycopg-binary>=3.2.7",
	"psycopg2-binary>=2.9.10",
	"python-dotenv>=1.1.0",
	"requests>=2.32.3",
	"sqlalchemy[asyncio]>=2.0.40",
	"uvicorn>=0.34.2",
	"whitenoise>=6.9.0",
	"xlrd>=2.0.1",
]

[dependency-groups]
dev = [
	"mypy>=1.15.0",
	"ruff>=0.11.6",
]
